Refactored the code for the linked lever and linked button - `1.20.1`
