+++ src/main/java/net/liukrast/eg/content/logistics/LinkedControlFrequencySlot.java
@@ -12,9 +12,9 @@ import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.Vec3;
 
-public class LinkedLeverFrequencySlot extends ValueBoxTransform.Dual {
+public class LinkedControlFrequencySlot extends ValueBoxTransform.Dual {
 
-    public LinkedLeverFrequencySlot(boolean first) {
+    public LinkedControlFrequencySlot(boolean first) {
         super(first);
     }
 
@@ -24,14 +24,14 @@ public class LinkedLeverFrequencySlot extends ValueBoxTransform.Dual {
 
         if(facing.getAxis().isHorizontal()) {
             return rotateHorizontally(state, isFirst()
-                    ? VecHelper.voxelSpace(6.2f, 4.5f, 3.20f)
-                    : VecHelper.voxelSpace(9.8f, 4.5f, 3.20f)
+                    ? VecHelper.voxelSpace(6.2f, 4.5f, 3.125f)
+                    : VecHelper.voxelSpace(9.8f, 4.5f, 3.125f)
             );
         }
 
         return VecHelper.rotateCentered(isFirst()
-                ? VecHelper.voxelSpace(9.8f, 3.5f, 4.4f)
-                : VecHelper.voxelSpace(6.1f, 3.5f, 4.4f)
+                ? VecHelper.voxelSpace(9.8f, 3.15f, 4.5f)
+                : VecHelper.voxelSpace(6.2f, 3.15f, 4.5f)
                 , facing == Direction.DOWN ? 180 : 0, Direction.Axis.X);
     }
