+++ src/main/java/net/liukrast/eg/content/logistics/LinkedLeverBlock.java
@@ -1,84 +1,24 @@
 package net.liukrast.eg.content.logistics;
 
-import com.simibubi.create.AllShapes;
-import com.simibubi.create.foundation.block.IBE;
-import com.simibubi.create.foundation.block.WrenchableDirectionalBlock;
-import net.liukrast.eg.registry.EGBlockEntityTypes;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.context.BlockPlaceContext;
-import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.LevelReader;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
-import net.minecraft.world.level.pathfinder.PathComputationType;
+import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.BlockHitResult;
-import net.minecraft.world.phys.shapes.CollisionContext;
-import net.minecraft.world.phys.shapes.VoxelShape;
 import org.lwjgl.system.NonnullDefault;
 
 @NonnullDefault
-public class LinkedLeverBlock extends WrenchableDirectionalBlock implements IBE<LinkedLeverBlockEntity> {
-
-    public static final BooleanProperty POWERED = BlockStateProperties.POWERED;
-
+public class LinkedLeverBlock extends LinkedControlBlock {
     public LinkedLeverBlock(Properties properties) {
         super(properties);
-        registerDefaultState(defaultBlockState().setValue(POWERED, false));
     }
-
-    @Override
-    public void neighborChanged(BlockState state, Level level, BlockPos pos, Block neighborBlock, BlockPos fromPos, boolean movedByPiston) {
-        if (level.isClientSide)
-            return;
-
-        Direction blockFacing = state.getValue(FACING);
-        if (fromPos.equals(pos.relative(blockFacing.getOpposite()))) {
-            if (!canSurvive(state, level, pos)) {
-                level.destroyBlock(pos, true);
-            }
-        }
-    }
-
-    @Override
-    public void onPlace(BlockState state, Level worldIn, BlockPos pos, BlockState oldState, boolean isMoving) {
-        if (state.getBlock() == oldState.getBlock() || isMoving)
-            return;
-        withBlockEntityDo(worldIn, pos, be -> be.transmit(getPower(state)));
-    }
-
-    @Override
-    public void onRemove(BlockState pState, Level pLevel, BlockPos pPos, BlockState pNewState, boolean pMovedByPiston) {
-        IBE.onRemove(pState, pLevel, pPos, pNewState);
-    }
-
-    private int getPower(BlockState state) {
-        return state.getValue(POWERED) ? 15 : 0;
-    }
-
-    @Override
-    public int getDirectSignal(BlockState blockState, BlockGetter blockAccess, BlockPos pos, Direction side) {
-        if (side != blockState.getValue(FACING))
-            return 0;
-        return getSignal(blockState, blockAccess, pos, side);
-    }
-
-    @Override
-    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
-        builder.add(POWERED);
-        super.createBlockStateDefinition(builder);
-    }
-
 
     @Override
     public InteractionResult use(BlockState state, Level level, BlockPos pos, Player player, InteractionHand p_60507_, BlockHitResult p_60508_) {
@@ -86,51 +26,13 @@ public class LinkedLeverBlock extends WrenchableDirectionalBlock implements IBE<
         return onBlockEntityUse(level, pos, be -> {
             var cycled = state.cycle(POWERED);
             level.setBlock(pos, cycled, 3);
-            float f = cycled.getValue(POWERED) ? 0.6F : 0.5F;
+            boolean isPowered = cycled.getValue(POWERED);
+            float f = isPowered ? 0.6F : 0.5F;
             level.playSound(player, pos, SoundEvents.LEVER_CLICK, SoundSource.BLOCKS, 0.3F, f);
+            level.gameEvent(player, isPowered ? GameEvent.BLOCK_ACTIVATE : GameEvent.BLOCK_DEACTIVATE, pos);
             be.transmit(getPower(cycled));
             return InteractionResult.SUCCESS;
         });
-    }
-
-    @Override
-    public BlockState getRotatedBlockState(BlockState originalState, Direction _targetedFace) {
-        return originalState;
-    }
-
-    @Override
-    public boolean canSurvive(BlockState state, LevelReader worldIn, BlockPos pos) {
-        BlockPos neighbourPos = pos.relative(state.getValue(FACING)
-                .getOpposite());
-        BlockState neighbour = worldIn.getBlockState(neighbourPos);
-        return !neighbour.canBeReplaced();
-    }
-
-    @Override
-    public BlockState getStateForPlacement(BlockPlaceContext context) {
-        BlockState state = defaultBlockState();
-        state = state.setValue(FACING, context.getClickedFace());
-        return state;
-    }
-
-    @Override
-    public VoxelShape getShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context) {
-        return AllShapes.LOGISTICAL_CONTROLLER.get(state.getValue(FACING));
-    }
-
-    @Override
-    public boolean isPathfindable(BlockState p_60475_, BlockGetter p_60476_, BlockPos p_60477_, PathComputationType p_60478_) {
-        return false;
-    }
-
-    @Override
-    public Class<LinkedLeverBlockEntity> getBlockEntityClass() {
-        return LinkedLeverBlockEntity.class;
-    }
-
-    @Override
-    public BlockEntityType<? extends LinkedLeverBlockEntity> getBlockEntityType() {
-        return EGBlockEntityTypes.LINKED_LEVER.get();
     }
 
 }
