+++ src/main/resources/assets/extra_gauges/models/block/linked_button/off.json
@@ -116,6 +116,8 @@
 			"scale": [0.625, 0.625, 0.625]
 		},
 		"fixed": {
+			"rotation": [270, 90, 0],
+			"translation": [0, 0, -4],
 			"scale": [0.5, 0.5, 0.5]
 		}
 	}
