+++ src/main/resources/assets/extra_gauges/models/block/linked_lever/off.json
@@ -129,6 +129,8 @@
 			"scale": [0.625, 0.625, 0.625]
 		},
 		"fixed": {
+			"rotation": [270, 90, 0],
+			"translation": [0, 0, -4],
 			"scale": [0.5, 0.5, 0.5]
 		}
 	}
